self-correcting-multi-agent-system/requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0

# Jupyter and visualization
jupyter>=1.0.0
jupyterlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0

# Web search and data retrieval
tavily-python>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Vector database and embeddings
pinecone-client>=2.2.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Database and data processing
# sqlite3
sqlalchemy>=2.0.0
datasets>=2.14.0

# Evaluation and metrics
scikit-learn>=1.3.0
rouge-score>=0.1.2
bert-score>=0.3.13

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
rich>=13.0.0
loguru>=0.7.0

# Code execution (optional)
jupyter-client>=8.0.0
ipykernel>=6.25.0

# Optional: For advanced features
transformers>=4.30.0
torch>=2.0.0